+++ src/test/java/com/baro/jwt_auth/user/controller/UserControllerTest.java
@@ -2,7 +2,7 @@ package com.baro.jwt_auth.user.controller;
 
 import com.baro.jwt_auth.common.exception.CustomException;
 import com.baro.jwt_auth.common.response.ErrorCode;
-import com.baro.jwt_auth.user.config.SecurityConfig;
+import com.baro.jwt_auth.config.SecurityConfig;
 import com.baro.jwt_auth.user.dto.request.LoginRequestDto;
 import com.baro.jwt_auth.user.dto.request.SignupRequestDto;
 import com.baro.jwt_auth.user.dto.response.LoginResponseDto;
@@ -10,8 +10,8 @@ import com.baro.jwt_auth.user.dto.response.SignupResponseDto;
 import com.baro.jwt_auth.user.dto.response.UserRoleUpdateResponseDto;
 import com.baro.jwt_auth.user.entity.UserEntity;
 import com.baro.jwt_auth.user.entity.UserRoleEnum;
-import com.baro.jwt_auth.user.jwt.JwtUtil;
-import com.baro.jwt_auth.user.security.UserDetailsServiceImpl;
+import com.baro.jwt_auth.security.jwt.JwtUtil;
+import com.baro.jwt_auth.security.UserDetailsServiceImpl;
 import com.baro.jwt_auth.user.service.UserService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.DisplayName;
